% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Hausaufgabe}
\alias{Hausaufgabe}
\title{Starte eine interaktive Hausaufgabe}
\usage{
Hausaufgabe(nr)
}
\arguments{
\item{nr}{Die Nummer der Hausaufgabe.}
}
\value{
Die interaktive Hausaufgabe wird gestartet.
}
\description{
Starte eine interaktive Hausaufgabe
}
\examples{
# starte die erste Hausaufgabe
Hausaufgabe(1)

}
